Added Node.js server
Files associated with the Node.js server:
- index.js
- js/router.js
- js/server.js
- js/requestHandlers.js

The following files are simply a WIP for a basic implementation of the course retrieval:
- html/index.html
- js/getUnits.js
At the moment they just send basic form data to the server. I left this in as an example for how to send such a request to the server.