Fixed bugs related to compiling the units in a plan's options
Bug 1: When checking for duplicates in plan with 2 majors, both
duplicate units would be removed rather than deleting the duplicate
and leaving just one. This was solved by looping through only the
units after the current unit when checking for duplicates.

Bug 2: Some majors (e.g., MJ-BUIS) were broken. This is because
there is an outer and an inner loop when checking for duplicates.
In the outer loop, electives would be ignored. However, the inner
loop was not checking for electives, therefore they would be
treated like a normal unit. The function would break when trying
to access the non-existent 'code' attribute in an elective.